Added an agent using a prioritized experience replay.
Also introduced a "tournament" script for making saved agents
battle against each other and record their stats.

Added an optional feature that makes the value-net and the
reference value-net change places periodically INSTEAD of
periodically assign the weights from the value-net to the
reference value-net. This was tried on a hunch, and the small-
scale preliminary test suggests it is good enough to keep
experimenting with.